requirements.txt
alembic==1.18.4
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.1
boto3==1.42.68
botocore==1.42.68
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gmpy2==2.3.0
google-play-scraper==1.2.7
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lightgbm==4.6.0
llvmlite==0.46.0
load-dotenv==0.1.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy==2.1.3
optree==0.19.0
optuna==4.8.0
optuna-integration==4.8.0
packaging==26.0
pandas==2.2.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
polars==1.39.0
polars-runtime-32==1.39.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pybind11==3.0.1
pybind11-global==3.0.1
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
stack_data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.14.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
xgboost==3.2.0
